// apps/humayro/lib/api/model/districtCreateDTO.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Humayro API
 * Humayro e-commerce API
 * OpenAPI spec version: 1.0
 */

export interface DistrictCreateDTO {
  /** @minLength 1 */
  name: string;
  nameUz?: string;
  regionId: number;
  fakturaDistrictCode?: number;
  fakturaDistrictName?: string;
  soato?: number;
}