// apps/humayro/lib/api/model/regionCreateDTO.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Humayro API
 * Humayro e-commerce API
 * OpenAPI spec version: 1.0
 */

export interface RegionCreateDTO {
  /** @minLength 1 */
  name: string;
  nameUz?: string;
  fakturaRegionId?: number;
  fakturaRegionCode?: number;
  fakturaRegionName?: string;
}